New method of specifying context: get_context()
This is much cleaner than shoehorning a new argument for all nodes, we
just define an extra optional method for Ops and have a bit of utility
code in Apply.

As a bonus, this also make context passing optional and will not add
any overhead for Ops that do not use a context.